+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @flowglad/shared
 
+## 0.2.4
+
+### Patch Changes
+
+- Add baseURL, use billing.retrieve
+
 ## 0.2.3
 
 ### Patch Changes